requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.27
django-cors-headers==4.9.0
django-filter==25.1
django-otp==1.6.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
google-auth==2.41.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
httplib2==0.31.0
idna==3.11
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.5
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
qrcode==8.2
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
